Cleaning up File Organization + taking out awkward floating shovel
changed the playerscript so that deflecting crabs depends on the boolean has_shovel rather than the shovel collider. Now the shovel gets destroyed so we can't depend on its collider to hit crabs.

Reorganized and changed a few names of our project folders so that explains why it says 1655 changed files loll