Immediate PICKUP after DROP fixed - wait 0.5seconds
temporary fix because I'm not 100% if disabling the user's box collider won't cause problems in the future, even if it is for half a second. I.e what if a seagull needs to detect the player's collider for half a second as they are grabbing an object but their collider is disabled? just a thought.